ch21

+++ 02.kernel64/src/console_shell.c
@@ -31,6 +31,8 @@ SHELLCOMMANDENTRY gs_vstCommandTable[] =
     {"killtask", "end task, ex)killtask 1(ID) or 0xffffffff(all task)", kKillTask},
     {"cpuload", "show processor load", kCPULoad},
     {"testmutex", "test mutex function", kTestMutex},
+    {"testthread", "Test Thread And Process Function", kTestThread},
+    {"showmatrix", "Show Matrix Screen", kShowMatrix},
 };
 
 // 셸의 메인 루프
@@ -436,7 +438,8 @@ static void kCreateTestTask(const char* pcParameterBuffer){
     // 타입 1 태스크 생성
     case 1:
         for(i=0; i<kAToI(vcCount, 10); i++){
-            if(kCreateTask(TASK_FLAGS_LOW, (qword)kTestTask1) == null){
+            if(kCreateTask(TASK_FLAGS_LOW | TASK_FLAGS_THREAD, 0, 0,
+                           (qword)kTestTask1) == null){
                 break;
             }
         }
@@ -447,7 +450,8 @@ static void kCreateTestTask(const char* pcParameterBuffer){
     case 2:
     default:
         for(i=0; i<kAToI(vcCount, 10); i++){
-            if(kCreateTask(TASK_FLAGS_LOW, (qword)kTestTask2) == null){
+            if(kCreateTask(TASK_FLAGS_LOW | TASK_FLAGS_THREAD, 0, 0,
+                           (qword)kTestTask2) == null){
                 break;
             }
         }
@@ -508,8 +512,11 @@ static void kShowTaskList(const char* pcParameterBuffer){
                 }
                 kPrintf("\n");
             }
-            kPrintf("[%d] Task ID[0x%q], Priority[%d], Flags[0x%q]\n", 1 + iCount++,
-                    pstTCB->stLink.qwID, GETPRIORITY(pstTCB->qwFlags), pstTCB->qwFlags);
+            kPrintf("[%d] Task ID[0x%q], Priority[%d], Flags[0x%q], Thread[%d]\n", 1 + iCount++,
+                    pstTCB->stLink.qwID, GETPRIORITY(pstTCB->qwFlags), pstTCB->qwFlags,
+                    kGetListCount(&(pstTCB->stChildThreadList)));
+            kPrintf("    Parent PID[0x%q], Memory Address[0x%q], Size[0x%q]\n",
+                    pstTCB->qwParentProcessID, pstTCB->pvMemoryAddress, pstTCB->qwMemorySize);
         }
     }
 }
@@ -535,21 +542,32 @@ static void kKillTask(const char* pcParameterBuffer){
 
     // 특정 ID만 종료하는 경우
     if(qwID != 0xffffffff){
-        kPrintf("kill task ID [0x%q] ", qwID);
-        if(kEndTask(qwID) == true){
-            kPrintf("success\n");
+        pstTCB = kGetTCBInTCBPool(GETTCBOFFSET(qwID));
+        qwID = pstTCB->stLink.qwID;
+
+        //시스템 테스트는 제외
+        if(((qwID >> 32) != 0) && ((pstTCB->qwFlags & TASK_FLAGS_SYSTEM) == 0x00)){
+            kPrintf("kill task ID [0x%q] ", qwID);
+            if(kEndTask(qwID) == true){
+                kPrintf("success\n");
+            }
+            else{
+                kPrintf("fail\n");
+            }
         }
         else{
-            kPrintf("fail\n");
+            kPrintf("task does not exists or task is system task\n");
         }
     }
 
     // 콘솔 셸과 유휴 태스크를 제외하고 모든 태스크 종료
     else{
-        for(int i=2; i<TASK_MAXCOUNT; i++){
+        for(int i=0; i<TASK_MAXCOUNT; i++){
             pstTCB = kGetTCBInTCBPool(i);
             qwID = pstTCB->stLink.qwID;
-            if((qwID >> 32) != 0){
+
+            // 시스템 태스크는 삭제 목록에서 제외
+            if(((qwID >> 32) != 0) && ((pstTCB->qwFlags & TASK_FLAGS_SYSTEM) == 0x00)){
                 kPrintf("kill task ID [0x%q]", qwID);
                 if(kEndTask(qwID) == true){
                     kPrintf("success\n");
@@ -613,8 +631,106 @@ static void kTestMutex(const char* pcParameterBuffer){
 
     for(i=0; i<3; i++){
         // 뮤텍스를 테스트하는 테스크를 3개 생성
-        kCreateTask(TASK_FLAGS_LOW, (qword)kPrintNumberTask);
+        kCreateTask(TASK_FLAGS_LOW | TASK_FLAGS_THREAD, 0, 0, (qword)kPrintNumberTask);
     }
     kPrintf("wait until %d task end...\n", i);
     kGetch();
+}
+
+// 태스크 2를 자신의 스레드로 생성하는 테스크
+static void kCreateThreadTask(){
+    for(int i=0; i<3; i++){
+        kCreateTask(TASK_FLAGS_LOW | TASK_FLAGS_THREAD, 0, 0, (qword)kTestTask2);
+    }
+    while(true){
+        kSleep(1);
+    }
+}
+
+// 스레드를 태스트하는 테스크 생성
+static void kTestThread(const char* pcParameterBuffer){
+    TCB* pstProcess;
+
+    pstProcess = kCreateTask(TASK_FLAGS_LOW | TASK_FLAGS_PROCESS, (void*)0xeeeeeeee, 0x1000,
+                             (qword)kCreateThreadTask);
+    if(pstProcess != null){
+        kPrintf("process [0x%q] create success\n", pstProcess->stLink.qwID);
+    }
+    else{
+        kPrintf("process create fail\n");
+    }
+}
+
+// 난수를 발생시키기 위한 변수
+static volatile qword gs_qwRandomValue = 0;
+
+// 임의의 난수를 반환
+qword kRandom(){
+    gs_qwRandomValue = (gs_qwRandomValue * 412153 + 5571031) >> 16;
+    return gs_qwRandomValue;
+}
+
+// 철자를 흘러내리게 하는 스레드
+static void kDropCharactorThread(){
+    int iX;
+    char vcText[2] = {0,};
+
+    iX = kRandom() % CONSOLE_WIDTH;
+
+    while(true){
+        // 잠시 대기함
+        kSleep(kRandom() % 20);
+
+        if((kRandom() % 20) < 16){
+            vcText[0] = ' ';
+            for(int i=0; i<CONSOLE_HEIGHT-1; i++){
+                kPrintStringXY(iX, i, vcText);
+                kSleep(50);
+            }
+        }
+        else{
+            for(int i=0; i<CONSOLE_HEIGHT-1; i++){
+                vcText[0] = i + kRandom();
+                kPrintStringXY(iX, i, vcText);
+                kSleep(50);
+            }
+        }
+    }
+}
+
+// 스레드를 생성하여 매트릭스 화면처럼 보여주는 프로세스
+static void kMatrixProcess(){
+    int i;
+    for(i=0; i<300; i++){
+        if(kCreateTask(TASK_FLAGS_THREAD | TASK_FLAGS_LOW, 0, 0,
+                        (qword)kDropCharactorThread) == null){
+            
+            break;
+        }
+
+        kSleep(kRandom() % 5 + 5);
+    }
+    kPrintf("%d thread is created\n", i);
+
+    // 키가 입력되면 프로세스 종료
+    kGetch();
+}
+
+//매트릭스 화면을 보여줌
+static void kShowMatrix(const char* pcParameterBuffer){
+    TCB* pstProcess;
+
+    pstProcess = kCreateTask(TASK_FLAGS_PROCESS | TASK_FLAGS_LOW, (void*)0xe00000, 0xe00000,
+                             (qword)kMatrixProcess);
+    if(pstProcess != null){
+        kPrintf("matrix process [0x%q] create success\n");
+
+        // 태스크 종료될 때까지 대기
+        while((pstProcess->stLink.qwID >> 32) != 0){
+            kSleep(100);
+        }
+    }
+    else{
+        kPrintf("matrix process create fail\n");
+    }
 }
